move test setup to test helper file
